Created a new Module to refine the Object, Proc and Symbol classes and updated the readme
