--- lib/wor/kernel.rb
@@ -1,9 +1,0 @@
-Kernel.class_eval do
-  def merge_method
-    raise Wor::Batchifier::Exceptions::InvalidStrategyType
-  end
-
-  def merge_base_case
-    raise Wor::Batchifier::Exceptions::InvalidStrategyType
-  end
-end
