--- lib/wor/proc.rb
@@ -1,9 +1,0 @@
-Proc.class_eval do
-  def merge_method
-    self
-  end
-
-  def merge_base_case
-    {}
-  end
-end
