+++ spec/wor/batchifier/batchifier_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
+using Wor::Batchifier
 
 describe Wor::Batchifier do
   describe 'execute_in_batches' do
@@ -126,15 +127,21 @@ describe Wor::Batchifier do
       end
 
       context 'when strategy is sent by block' do
-        let(:no_response_proc) { Proc.new { { } } }
+        let(:array_merge_proc) {
+          Proc.new do |response, rec|
+            rec = [] if rec.empty?
+            rec + response
+          end
+        }
+
         let(:batchified_process) do
-          execute_in_batches(ids,batch_size: 10,
-                                 strategy: no_response_proc) do |chunk|
+          execute_in_batches(ids, batch_size: 10,
+                                  strategy: array_merge_proc) do |chunk|
             chunk.map { |id| id + 10 }
           end
         end
 
-        let(:expected_response) { { } }
+        let(:expected_response) { ids.map { |id| id + 10 } }
 
         it 'returns the expected result' do
           expect(batchified_process).to eq(expected_response)
